+++ R/misc_funs.R
@@ -621,13 +621,12 @@ fsignif = signif_plus = function (x, s = 2, r = 0, commas = TRUE){
 
   if(commas) res = sapply(res, commas_single, s = s, r = r)
 
-  qui0 = grepl("^0.", res, perl = TRUE)
-  if(any(qui0)){
-    qui_short = nchar(res) < s + 2
-    if(any(qui_short)){
-      for(i in which(qui0)[qui_short]){
-        res[i] = as.vector(sprintf("%s%.*s", res[i], s + 2 - nchar(res[i]), "0000000000000000"))
-      }
+  qui0 = which(grepl("^0.", res, perl = TRUE))
+  if(length(qui0) > 0){
+    qui_short = nchar(res[qui0]) < s + 2
+    qui2fix = qui0[qui_short]
+    for(i in qui2fix){
+      res[i] = as.vector(sprintf("%s%.*s", res[i], s + 2 - nchar(res[i]), "0000000000000000"))
     }
   }